Use size_t for PME grid indexing and add DEBYE constant
- Change grid index calculations from int to size_t to prevent overflow
- Add static constexpr DEBYE constant to header
- Add clarifying comment for PCG early break case

+++ platforms/reference/src/ReferencePMETholeDipoleForce.h
@@ -77,7 +77,7 @@ private:
     Vec3 _recipBoxVectors[3];
     Vec3 _periodicBoxVectors[3];
 
-    int _totalGridSize;
+    size_t _totalGridSize;
     IntVec _pmeGridDimensions;
 
     std::vector<std::complex<double>> _pmeGrid;
